refactor: make component for duplicate codes between http and quiz
