+++ frontend/src/components/general/general.tsx
